extract storage
